requirements.txt
# Vision Language Model Support
openai>=1.12.0

# Web Framework
streamlit>=1.31.0

# Visualization
plotly>=5.18.0
pandas>=2.0.0

# Image Processing
Pillow>=10.0.0
pdf2image>=1.16.3

# Utilities
python-dotenv>=1.0.0